+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    as well as shortcut key sequences.
  - Colors for the audio scope and graph editors are now customizable
  - Autosave. Disabled by default, enable in the General tab in Config.
+ - Jxy effect for setting the global volume scale.
+ - Menu action in Tracker for reseting the global volume (Tracker|Reset global volume)
 ### Changed
  - Combined the Sound and Midi tabs in the Configuration dialog (Sound + MIDI)
  - AZERTY keyboard layout uses French AZERTY
@@ -22,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    the right of the pattern editor.
  - Pattern editor row numbers can be displayed in decimal (default is hexadecimal)
  - Cursor wrap can now be disabled in General settings
+ - Hxx effect (set sweep register) now works.
 ### Fixed
  - Issue with Instrument Editor not properly updating when changing instruments
  - Hitting enter in the sequence editor line edit opens the Wave editor